$(FrameworkPathOverride) property should respect $(TargetFrameworkRootPath)
`FrameworkPathOverride` set in `Microsoft.Common.CurrentVersion.targets`,
does not use the $(TargetFrameworkRootPath) while calling
ToolLocationHelper.GetPathToStandardLibraries(...) .

    <FrameworkPathOverride Condition="'$(FrameworkPathOverride)' == ''">$([Microsoft.Build.Utilities.ToolLocationHelper]::GetPathToStandardLibraries($(TargetFrameworkIdentifier), $(TargetFrameworkVersion), $(TargetFrameworkProfile), $(PlatformTarget)))</FrameworkPathOverride>

This means that if the project is targeting a custom framework,
which is installed in a directory different from the default root path,
then it won't be able to find the framework and eventually,
`FrameworkPathOverride` will be set to an incorrect path. This can be
seen with mono/osx which has an alternate directory (
`/Library/Frameworks/Mono.framework/External/xbuild-frameworks` ) for
Xamarin mobile frameworks.

This manifests as an `_ExplicitReference` getting added to
`mscorlib.dll` but with the wrong path, causing compilation to fail.