[RFC]: Add support for fallback search paths for Target frameworks
Currently, target frameworks are looked up in the default location or a
path specified by $(TargetFrameworkRootPath). It would be useful to be
able to look for these frameworks in more than one location. For
example, in case of Xamarin products on OSX, mobile specific frameworks
are installed in
`/Library/Frameworks/Mono.framework/External/xbuild-frameworks` and the
.NET ones are installed in
`/Library/Frameworks/Mono.framework/Versions/Current/lib/mono/xbuild-frameworks`
.

With support for fallback search paths, a lookup for a target framework will follow the order:

    1. $(TargetFrameworkRootPath) or Default location is $(TargetFrameworkRootPath) is ''
    2. Fallback search paths

These fallback search paths can be set in app.config, per os, like:

<msbuildToolsets default="14.1">
  <toolset toolsVersion="14.1">
    <property name="TargetFrameworkRootPathSearchPathsOSX" value="/tmp/foo;/tmp/bar" />
    <property name="TargetFrameworkRootPathSearchPathsWindows" value="C:\tmp\foo;C:\tmp\bar" />
    ...

This patch is not complete yet - tests need to be added.