[tmp] MSBuildFrameworkToolsPath should be point to the framework tools
.. directory and not to the msbuild tools directory. In xplat, this is
being set to a Windows-specific path in app.config, but for this branch
we don't have that and so need to depend on the value set in the code.

This manifested as targets like `*WinFx.targets` getting imported twice,
once from `MSBuildToolsPath` and then from `MSBuildFrameworkToolsPath`,
which were ending up as the same directory and so a warning was seen.