Workaround roslyn not correctly handling `/debug+` on OSX
On OSX, roslyn required `/debug:portable` earlier to generate a portable
pdb. But it was fixed to map `/debug:{full,pdbonly}` to
`/debug:portable`.

But this only works if the above argument is explicitly specified. OTOH,
if only `/debug+` is used, then the mapping is not performed, causing an
error like:

CSC : error CS0041: Unexpected error writing debug information -- 'Windows PDB writer is not available -- could not find Microsoft.DiaSymReader.Native.amd64.dll'

Also, we want to do this only only if required (eg. for Debug configuration).