Exec codepage manipulation in .NET Core (#2231)
Fixes #2187 by fixing the broken assumption that .NET Core MSBuild is
always running under a Unicode-compatible codepage. That's not true on
.NET Core on Windows, for the same reason the original checks were
needed on full framework on Windows.

Prior to this change, .NET Core MSBuild on Windows behaved like .NET
Core MSBuild on other OSes and treated its output as UTF-8. But that's
not true on Windows because of OS limitations, not .NET ones. This
commit brings back the logic used for full framework MSBuild (except the
unavailable is-Windows-too-old check).