[RFC][mono] Add suport for denied assemblies, fixes bxc#56682
Mono denies loading certain "bad" assemblies because they include an
implementation that doesn't work under mono. Instead, mono provides it's
own implementations of the same which are safe to load.

These are specific versions from specific nugets. Mono installs a list
of these in $(MSBuildExtensionsPath)/deniedAssembliesList.txt .

If these bad assemblies get bundled with an app and the runtime tries to
load them, then they will get denied. If this is running from a full
mono installation, then the runtime continues looking for the assembly
in other locations.

But if this is a case, like a Mac app bundle (with Xamarin.Mac), which
uses it's own bundled mono+assemblies, then the runtime won't be able to
fallback to anything else. So, we want to ensure that we bundle the
"good" assemblies instead.

This is not an issue for XA/XI as they have their own
TargetFrameworkMoniker and so the nuget has `_._` for them. IOW, msbuild
uses the framework assemblies there.

Solution:

- We tried an approach of: filtering out such references
  (`@(Reference)`) before the `ResolveAssemblyReference` task. All the
  facades are passed to the compiler as references, so the correct
  assemblies will get pulled in accordingly.

  But the problem with this was that if after removing these, we ended
  up with no facade references, then `ResolveAssemblyReference` returns
  `DependsOnSystemRuntime=false`, which causes the facade assemblies to
  not be implicitly passed to the compiler!

- Instead, now we get the list of target framework directories and
  facade directories and look for the "good" assemblies in there. If we
  can find the assembly, then we update the reference (ItemSpec or HintPath) to
  the new path.
  Else, we warn the user and don't touch the reference!

For this, we add a `FilterDeniedAssemblies` tasks in a new
`Mono.Build.Tasks.dll`.

Note: the list is part of this commit for now. Before the merge, this
will be generated+installed from mono instead.