[mono] Filter windows-specific net4x assemblies
We bundle net4x assemblies from Microsoft.NET.Build.Extensions, some of
which are windows-specific and not compatible with mono/!windows. We
have an existing mechanism `FilterDeniedAssemblies` to filter out such
assemblies before reference resolution.

This was enabled only by Xamarin.Mac . But here we enable it for any
.net framework project (4.6.1 - 4.7) which might directly or indirectly
reference netstandard, since in this case the bundled net4x assemblies
get used.

Fixes: #60315