Fix tests failing in Release mode only
Traits class caches the values of various environment variables by
default. But when built in Debug configuration, it allows resetting the
cache through another environment variable.

There are some tests that depend on being able to reset the cache and
then changing the value of various environment variables and these fail
in Release configuration:

      Microsoft.Build.UnitTests.Evaluation.ItemEvaluation_Tests.LazyWildcardExpansionDoesNotEvaluateWildCardsIfNotReferenced [FAIL]
      Microsoft.Build.Engine.UnitTests.BackEnd.SdkResolverLoader_Tests.VerifyThrowsWhenResolverHasNoPublicConstructor [FAIL]
      Microsoft.Build.UnitTests.Evaluation.Evaluator_Tests.ItemPredecessorToItemWithCaseChangeAndEscapeHatch [FAIL]
      Microsoft.Build.UnitTests.Evaluation.Evaluator_Tests.ItemPredecessorToItemViaTransformWithCaseChangeWithEscapeHatch [FAIL]
      Microsoft.Build.UnitTests.Evaluation.Expander_Tests.PropertyStaticFunctionAllEnabled [FAIL]
      Microsoft.Build.UnitTests.OM.Definition.Project_Tests.ProjectImportedEventMissingFileNoGlobMatch [FAIL]

So, We enable it for all builds. The env var
(`MSBUILDRELOADTRAITSONEACHACCESS`) gets checked only if we are running
under tests.