[mono] Fix unintended use of shared compiler server
The previous commit which removed the `$(CscToolPath)` override, has the
side effect that the recent roslyn versions (2.7.x?) will use the shared
compiler, which doesn't yet work well on mono.

Earlier, `$(UseSharedCompilation)` was defaulting to true but it didn't
matter because roslyn didn't support the compiler server anyway.

Now it seems to work on netcore, but not (well enough) on mono. So, it
uses the compiler server if (essentially):

	1. `$(UseSharedCompilation)` is true (default)
	2. `$(CscToolPath)` is not overridden.

Since, the previous patch makes (2) true, we end up with the compiler
server being used. So, explicitly disable that for now.