Use NuGetPackageRoot in setting NuGetPackagesDir for common.props
This patch may serve to prevent a spurious error message under some
build configurations, during the restore MSBuild target for the MSBuild
project file ${WRKSRC}/mono/build/update_bundled_bits.proj

 "Tried to restore, but seems to have failed for ..."

This error message would occur from the Mono MSBuild project support file
 ${WRKSRC}/mono/build/sdks_and_nugets/deploy_from_nugets.proj
as under the MSBuild target, EnsureDependencyRestored, subsequent to
evaluation of the project support file
 ${WRKSRC}/mono/build/sdks_and_nugets/update_sdks_and_nugets.proj
namely during MSBuild evaluation of the top-level project file
 ${WRKSRC}/ mono/build/update_bundled_bits.proj

This error would occur when a *.nuspec file does not exist at a specific
pathname -- that pathname, computed as extensional to the local value of
the parameter, NuGetPackagesDir -- subsequent to the restore task under
the bootstrap msbuild environment, with update_bundled_bits.proj

The error may occur even after a successful restore, if the pathname
used in the corresponding EnsureDependencyRestored target does not match
the pathname used in the restore procedure.

The following patch endeavors to initialize the value of the
NuGetPackagesDir parameter, conditionally, with a pathname that may be
interoperable with the pathname semantics in the msbuild restore.

Subsequently, the build for mono/build/update_bundled_bits.proj may be
produced with an MSBUild parameter
  /p:NuGetPackageRoot=${MSBUILD_SRC_ROOT}/.packages

The pathname semantics applied here is roughly analogous to that used
in Microsoft.DotNet.Arcade.Sdk version 1.0.0-beta.19461.7, nupkg
resource file tools/RepoLayout.props

Additional Remarks

* The msbuild restore target may exit abnormally, per default, if
  the restore task has failed. As such, the EnsureDependencyRestored
  MSBuild target may seem redundant, in some regards.

* The NormalizeDirectory opern, as used below, adds a trailing slash to
  the string, such that may be redundant in uses of the task
  EnsureDependencyRestored

* Patch developed for mono 0.08 from local pkgsrc devo (thinkum-sys)